+++ Frends.SFTP.WriteFile/Frends.SFTP.WriteFile/Definitions/Result.cs
@@ -1,7 +1,5 @@
-﻿using System.ComponentModel;
+﻿using Renci.SshNet.Sftp;
 using System.ComponentModel.DataAnnotations;
-
-#pragma warning disable 1591
 
 namespace Frends.SFTP.WriteFile.Definitions
 {
@@ -11,34 +9,22 @@ namespace Frends.SFTP.WriteFile.Definitions
     public class Result
     {
         /// <summary>
-        /// The name of the file. Does not include the path.
+        /// Full path to the written file.
         /// </summary>
+        /// <example>/destination/newfile.txt</example>
         [DisplayFormat(DataFormatString = "Text")]
-	    public string FileName { get; private set; }
+	    public string Path { get; private set; }
 
         /// <summary>
-        /// The full source path of the file.
+        /// Size of the new file in destination.
         /// </summary>
-        [DisplayFormat(DataFormatString = "Text")]
-        public string SourcePath { get; private set; }
+        /// <example>3.2</example>
+        public double SizeInMegaBytes { get; private set; }
 
-        /// <summary>
-        /// The destination path
-        /// </summary>
-        [DisplayFormat(DataFormatString = "Text")]
-        public string DestinationPath { get; private set; }
-
-        /// <summary>
-        /// Boolean value of the successful transfer.
-        /// </summary>
-        public bool Success { get; private set; }
-
-        public Result(string name, string sourcePath, string destinationPath, bool success)
+        internal Result(SftpFile file)
         {
-            FileName = name;
-            SourcePath = sourcePath;
-            DestinationPath = destinationPath;
-            Success = success;
+            Path = file.FullName;
+            SizeInMegaBytes = file.Length / 1024 / 1024;
         }
     }
 }