+++ Frends.SFTP.WriteFile/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.1] - 2022-06-14
+### Changed
+- Changed the main method to write from stream to a file.
+- Updated tests
+- Updated Renci.SshNet library
+
 ## [1.0.0] - 2022-03-10
 ### Added
 - Initial implementation